Allow choice of rotation; clockwise/anticlockwise
Fixes #105

Creates the `rotation` option that allows a setting of `anticlockwise`
(or 'acw') to reverse the direction of progression used by the knob. It
fully supports the clicking/touching and `displayPrevious`
functionality.

To facilitate extending this behavior to custom draw functions, arc
calculations have been factored out of `draw` into an `arc` function,
which returns the start and end angles, and a boolean denoting the
direction of the stroke.