Numeric validation for the colored foreground arc
It is possible for the data value to include non-numeric characters. This happens explicitly with modifications such as: http://stackoverflow.com/questions/18955708/how-to-display-units-with-jquery-knob

This modification works well for the label, but the addition of the '%' character causes the arc not to be displayed when the page is refreshed. The above fixes this and removes all non-number characters from the value before attempting to plot it.