Add an optionnal circle knob
Allows to optionally have a knob above the stroke.
It reduces the whole widget radius if activated, or the knob would overflow the canvas.

The knob's own radius must be expressed as a percentage of the ```lineWidth``` ( 0 < r < 1 ).

If no ```knobRadius``` config argument is given, the value is set to ```0``` to avoid adding special cases further in the code for users that don't use it.