added update event listener
bound a listener for an `update` event which calls the `val` method with the second param as `false` so as not to trigger the release handler.
this is useful when we create a knob that we update periodically as well expect the user to interact with at the same time e.g. an audio scrubber which shows the track progress which you can change as well.